move mockRunner into dedicated file
will be used also elsewhere. Does not make sense to have that in one of the specific test files
when it is shared between all the tests.

Disadvantage (maybe someone knows how to improve): mock.go does not have test scope, the
content is also available outside tests.
